JAMES-4077 - Introduce james-server-guice-lucene module & replace it to JPAJamesServerMain
