+++ server/apps/jpa-app/pom.xml
@@ -110,6 +110,10 @@
         </dependency>
         <dependency>
             <groupId>${james.groupId}</groupId>
+            <artifactId>james-server-guice-lucene</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>${james.groupId}</groupId>
             <artifactId>james-server-guice-mailbox</artifactId>
         </dependency>
         <dependency>
